Extract FileWrappers logic and reuse it for saving af UIKit
